5ZCMF7C fold sync into api/state and the ticket helpers into api/issue-helpers
